gui: Update deprecated calls to wxFont
In 2014, a deprecation was announced changing how wxFont is called to
ensure stronger typing (http://trac.wxwidgets.org/ticket/15814).  This
change was introduced in version 2.9.5 of wxWidgets.

Compiling with a recent version of GCC (10.2.1) surfaces these warnings
as errors due to the flag ``.

The error presented is:

> ‘wxFont::wxFont(int, int, int, int, bool, const wxString&,
wxFontEncoding)’ is deprecated: use wxFONT{FAMILY,STYLE,WEIGHT}_XXX
constants

Additionally, the documentation now suggests the use of a helper class
(`wxFontInfo`) instead of directly calling the wxFont constructor.

https://docs.wxwidgets.org/3.1/classwx_font_info.html